# 3o Semestre - Missão Certificação 3/api-docs_open_api.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/pessoas/{id}":{"put":{"tags":["pessoa-controller"],"operationId":"atualizar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PessoaAtualizacaoDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/notificacoes/{id}":{"put":{"tags":["notificacao-controller"],"operationId":"submeter","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/integracao/ecartas/{id}":{"put":{"tags":["integracao-controller"],"operationId":"atualizarCarta","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/integracao/dje/{id}":{"put":{"tags":["integracao-controller"],"operationId":"atualizarDje","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/pessoas":{"get":{"tags":["pessoa-controller"],"operationId":"listar","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Pessoa"}}}}}}},"post":{"tags":["pessoa-controller"],"operationId":"criar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PessoaCriacaoDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Pessoa"}}}}}}},"/notificacoes":{"get":{"tags":["notificacao-controller"],"operationId":"listar_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Notificacao"}}}}}}},"post":{"tags":["notificacao-controller"],"operationId":"criar_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotificacaoCriacaoDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/integracao/ecartas":{"get":{"tags":["integracao-controller"],"operationId":"listarCartas","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Notificacao"}}}}}}}},"/integracao/dje":{"get":{"tags":["integracao-controller"],"operationId":"listarDje","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Notificacao"}}}}}}}}},"components":{"schemas":{"PessoaAtualizacaoDTO":{"type":"object","properties":{"nome":{"type":"string"},"cep":{"type":"string"},"numeroEndereco":{"type":"string"},"complemento":{"type":"string"},"email":{"type":"string"}}},"PessoaCriacaoDTO":{"type":"object","properties":{"nome":{"type":"string"},"documento":{"type":"string"},"cep":{"type":"string"},"numeroEndereco":{"type":"string"},"complemento":{"type":"string"},"email":{"type":"string"}}},"Endereco":{"required":["cep"],"type":"object","properties":{"id":{"type":"integer","format":"int32"},"cep":{"type":"string"},"logradouro":{"type":"string"},"bairro":{"type":"string"},"localidade":{"type":"string"},"uf":{"type":"string"}}},"Pessoa":{"required":["documento","nome"],"type":"object","properties":{"id":{"type":"integer","format":"int32"},"nome":{"type":"string"},"documento":{"pattern":"\\d{11}|\\d{14}","type":"string"},"endereco":{"$ref":"#/components/schemas/Endereco"},"numeroEndereco":{"type":"string"},"complemento":{"type":"string"},"email":{"type":"string"}}},"NotificacaoCriacaoDTO":{"type":"object","properties":{"documento":{"type":"string"},"numProcesso":{"type":"string"},"tipoDocumento":{"type":"string","enum":["IntimacaoJudicial","NotificacaoProcesso","ConclusaoProcesso","DecisaoJudicial","PagamentoSentenca","PagamentoCustas","Outros"]},"descricao":{"type":"string"}}},"Notificacao":{"required":["dataCadastro","descricao","numProcesso","pessoa","status","tipoDocumento"],"type":"object","properties":{"id":{"type":"integer","format":"int32"},"pessoa":{"$ref":"#/components/schemas/Pessoa"},"numProcesso":{"pattern":"^\\d{7}-\\d{2}.\\d{4}.\\d.\\d{2}.\\d{4}$","type":"string"},"tipoDocumento":{"type":"string","enum":["IntimacaoJudicial","NotificacaoProcesso","ConclusaoProcesso","DecisaoJudicial","PagamentoSentenca","PagamentoCustas","Outros"]},"status":{"type":"string","enum":["Rascunho","Submetida","Enviada","Recebida"]},"tipoNotificacao":{"type":"string","enum":["ECartas","Email","DJe"]},"descricao":{"type":"string"},"dataCadastro":{"type":"string","format":"date-time"},"dataSubmissao":{"type":"string","format":"date-time"},"dataEnvio":{"type":"string","format":"date-time"},"dataRecebimento":{"type":"string","format":"date-time"}}}}}}