# 3o Semestre - Missão Certificação 3/src/main/java/com/notificacoes/pje/model/Endereco.java
package com.notificacoes.pje.model;

import jakarta.persistence.Column;
import jakarta.persistence.Entity;
import jakarta.persistence.GeneratedValue;
import jakarta.persistence.GenerationType;
import jakarta.persistence.Id;
import jakarta.validation.constraints.NotBlank;
import lombok.Getter;
import lombok.Setter;

@Entity
@Getter
@Setter
public class Endereco {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Integer id;

    @NotBlank
    @Column(nullable = false, unique = true)
    private String cep;

    private String logradouro;
    private String bairro;
    private String localidade;
    private String uf;
}